requirements.txt
pydigitemp
pymongo
requests
pytz
